Added full AzureAD and endpoint authorization

+++ Winter Classes App/Controllers/CompaniesController.cs
@@ -3,23 +3,25 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 using Winter_Classes_App.EntityFramework;
 using Winter_Classes_App.Models;
 
 namespace Winter_Classes_App.Controllers
 {
-    public class CompaniesController : Controller
+    public class CompaniesController : BaseController
     {
-        private readonly DataContext _context;
 
-        public CompaniesController(DataContext context)
-        {
-            _context = context;
-        }
+        public CompaniesController(IConfiguration Configuration, DataContext context) : base(Configuration, context) { }
 
         public async Task<IActionResult> Index([FromQuery(Name = "search")] string searchString)
         {
-
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.HR)
+            {
+                return NotFound();
+            }
             if (String.IsNullOrEmpty(searchString))
             {
                 return View(await _context.Companies.ToListAsync());
@@ -34,8 +36,14 @@ namespace Winter_Classes_App.Controllers
 
         }
 
-        public IActionResult Create()
+        public async Task<IActionResult> CreateAsync()
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.ADMIN)
+            {
+                return NotFound();
+            }
             return View();
         }
 
@@ -43,6 +51,12 @@ namespace Winter_Classes_App.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name")] Company company)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.ADMIN)
+            {
+                return NotFound();
+            }
             if (ModelState.IsValid)
             {
                 _context.Add(company);
@@ -54,6 +68,12 @@ namespace Winter_Classes_App.Controllers
 
         public async Task<IActionResult> Edit(int? id)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.ADMIN)
+            {
+                return NotFound();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -71,6 +91,12 @@ namespace Winter_Classes_App.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] Company company)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.ADMIN)
+            {
+                return NotFound();
+            }
             if (id != company.Id)
             {
                 return NotFound();
@@ -101,6 +127,12 @@ namespace Winter_Classes_App.Controllers
 
         public async Task<IActionResult> Delete(int? id)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.ADMIN)
+            {
+                return NotFound();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -120,6 +152,12 @@ namespace Winter_Classes_App.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.ADMIN)
+            {
+                return NotFound();
+            }
             var company = await _context.Companies.FindAsync(id);
             _context.Companies.Remove(company);
             await _context.SaveChangesAsync();
