+++ Winter Classes App/Controllers/JobOffersController.cs
@@ -1,26 +1,22 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 using Winter_Classes_App.EntityFramework;
 using Winter_Classes_App.Models;
 
 namespace Winter_Classes_App.Controllers
 {
-    public class JobOffersController : Controller
+    public class JobOffersController : BaseController
     {
-        private readonly DataContext _context;
+        public JobOffersController(IConfiguration Configuration, DataContext context) : base(Configuration, context) { }
 
-        public JobOffersController(DataContext context)
-        {
-            _context = context;
-        }
         public async Task<IActionResult> Index([FromQuery(Name = "search")] string searchString) {
-
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
             if (String.IsNullOrEmpty(searchString))
             {
                 return View(await _context.JobOffers
@@ -40,6 +36,9 @@ namespace Winter_Classes_App.Controllers
 
         public async Task<IActionResult> Details(int? id)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+
             if (id == null)
             {
                 return NotFound();
@@ -56,8 +55,14 @@ namespace Winter_Classes_App.Controllers
             return View(jobOffer);
         }
 
-        public IActionResult Create()
+        public async Task<IActionResult> CreateAsync()
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel != PrivilegesLevel.LOGGEDIN)
+            {
+                return NotFound();
+            }
             var model = new JobOfferCreateView
             {
                 Companies = _context.Companies
@@ -69,6 +74,12 @@ namespace Winter_Classes_App.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(JobOfferCreateView model)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel != PrivilegesLevel.LOGGEDIN)
+            {
+                return NotFound();
+            }
             if (!ModelState.IsValid)
             {
                 model.Companies = _context.Companies;
@@ -94,6 +105,12 @@ namespace Winter_Classes_App.Controllers
 
         public async Task<IActionResult> Edit(int? id)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.HR)
+            {
+                return NotFound();
+            }
             if (id == null)
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             var jobOffer = await _context.JobOffers.Include(j => j.Company).FirstOrDefaultAsync(m => m.Id == id);
@@ -105,6 +122,12 @@ namespace Winter_Classes_App.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(JobOffer model)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.HR)
+            {
+                return NotFound();
+            }
             if (!ModelState.IsValid) return View();
             var jobOffer = await _context.JobOffers.FindAsync(model.Id);
             jobOffer.JobTitle = model.JobTitle;
@@ -133,6 +156,12 @@ namespace Winter_Classes_App.Controllers
 
         public async Task<IActionResult> Delete(int? id)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.HR)
+            {
+                return NotFound();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -153,6 +182,12 @@ namespace Winter_Classes_App.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            PrivilegesLevel privilegesLevel = await CheckGroup();
+            ViewBag.PrivilegesLevel = (int)privilegesLevel;
+            if (privilegesLevel < PrivilegesLevel.HR)
+            {
+                return NotFound();
+            }
             var jobOffer = await _context.JobOffers.FindAsync(id);
             _context.JobOffers.Remove(jobOffer);
             await _context.SaveChangesAsync();
